Финальное задание 12 го спринта. Расширяем функционал приложения. Доработана модель данных для дальнейшего развития проекта (хранение данных в базе SQL)
подключена база H2

+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -30,12 +30,12 @@ public class FilmController {
     }
 
     @PutMapping("/{filmId}/like/{userId}")
-    public void addLike(@PathVariable int filmId, @PathVariable int userId) {
+    public void addLike(@PathVariable long filmId, @PathVariable long userId) {
         filmService.addLike(filmId, userId);
     }
 
     @DeleteMapping("/{filmId}/like/{userId}")
-    public void removeLike(@PathVariable int filmId, @PathVariable int userId) {
+    public void removeLike(@PathVariable long filmId, @PathVariable long userId) {
         filmService.removeLike(filmId, userId);
     }
