+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -30,22 +30,22 @@ public class UserController {
     }
 
     @PutMapping("/{userId}/friends/{friendId}")
-    public void addFriend(@PathVariable int userId, @PathVariable int friendId) {
+    public void addFriend(@PathVariable Long userId, @PathVariable Long friendId) {
         userService.addFriend(userId, friendId);
     }
 
     @DeleteMapping("/{userId}/friends/{friendId}")
-    public void removeFriend(@PathVariable int userId, @PathVariable int friendId) {
+    public void removeFriend(@PathVariable Long userId, @PathVariable Long friendId) {
         userService.removeFriend(userId, friendId);
     }
 
     @GetMapping("/{userId}/friends")
-    public Set<User> getUserFriends(@PathVariable int userId) {
+    public Set<User> getUserFriends(@PathVariable Long userId) {
         return userService.getUserFriends(userId);
     }
 
     @GetMapping("/{userId1}/friends/common/{userId2}")
-    public List<User> getCommonFriends(@PathVariable int userId1, @PathVariable int userId2) {
+    public List<User> getCommonFriends(@PathVariable Long userId1, @PathVariable Long userId2) {
         return userService.getCommonFriends(userId1, userId2);
     }
 }
