+++ src/main/java/ru/yandex/practicum/filmorate/model/Friendship.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class Friendship {
 
-    private int userId;      // ID пользователя, который отправил запрос на дружбу
-    private int friendId;    // ID пользователя, которому был отправлен запрос
+    private long userId;      // ID пользователя, который отправил запрос на дружбу
+    private long friendId;    // ID пользователя, которому был отправлен запрос
     private FriendshipStatus status;  // Статус дружбы (подтверждённая или неподтверждённая)
 }
