+++ src/main/java/ru/yandex/practicum/filmorate/storage/InMemoryUserStorage.java
@@ -14,8 +14,8 @@ import java.util.Optional;
 @Component
 public class InMemoryUserStorage implements UserStorage {
 
-    private final Map<Integer, User> users = new HashMap<>();
-    private int nextId = 1;
+    private final Map<Long, User> users = new HashMap<>();
+    private long nextId = 1;
 
     @Override
     public User addUser(User user) {
@@ -42,7 +42,7 @@ public class InMemoryUserStorage implements UserStorage {
     }
 
     @Override
-    public Optional<User> getUserById(int id) {
+    public Optional<User> getUserById(long id) {
         return Optional.ofNullable(users.get(id));
     }
 
@@ -52,7 +52,7 @@ public class InMemoryUserStorage implements UserStorage {
     }
 
     @Override
-    public void deleteUser(int id) {
+    public void deleteUser(long id) {
         users.remove(id);
     }
